requirements.txt
PyYAML==3.12
requests==2.17.3